// src/core/purchase/enum/purchase.enum.ts
export enum PURCHASE_ERROR {
  PURCHASE_NOT_FOUND = 'PURCHASE_NOT_FOUND',
  AUTH_CODE_IS_INCORRECT = 'AUTH_CODE_IS_INCORRECT',
  COUNT_GREATER_MAXIMUM = 'COUNT_GREATER_MAXIMUM',
  LENGTH_GREATER_MAXIMUM = 'LENGTH_GREATER_MAXIMUM',
  MINIMUM_COUNT_IS = 'MINIMUM_COUNT_IS',
  MINIMUM_LENGTH_IS = 'MINIMUM_LENGTH_IS',
  PRODUCT_WAS_PURCHASED = 'PRODUCT_WAS_PURCHASED',
}